requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-jazzmin==2.6.0
django-rest-swagger==2.2.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
environs==9.5.0
Faker==18.10.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
boto3
s3fs
pre-commit==3.3.3
autoflake==2.2.0
black==23.7.0
isort==5.12.0
social-auth-core==4.4.2
